+++ pytest_cases/tests/fixtures/test_issue_github_54.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pytest_cases.main_fixtures import InvalidParamsList
-from pytest_cases import pytest_parametrize_plus, fixture_ref
+from pytest_cases import parametrize_plus, fixture_ref
 
 
 @pytest.fixture
@@ -11,6 +11,6 @@ def test():
 
 def test_invalid_argvalues():
     with pytest.raises(InvalidParamsList):
-        @pytest_parametrize_plus('main_msg', fixture_ref(test))
+        @parametrize_plus('main_msg', fixture_ref(test))
         def test_prints(main_msg):
             print(main_msg)